+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ---
 
 ## master
+* Upgrade kotlin to 1.5.31 in JsPlugin
+* MockingBird plugin
 
 ## 1.15.0
 * Support for iosSimulatorArm64